Shelley.Spec.NonIntegral

Documentation

(***) ∷ (RealFrac a, Enum a, Show a) ⇒ a → a → a

Exponentiation

exp' ∷ (RealFrac a, Show a) ⇒ a → a

ln' ∷ (RealFrac a, Enum a, Show a) ⇒ a → a

Compute natural logarithm via continued fraction, first splitting integral part and then using continued fractions approximation for `ln(1+x)`

findE ∷ RealFrac a ⇒ a → a → Integer

find n with `e^n<=x<e^(n+1)`

splitLn ∷ (RealFrac a, Show a) ⇒ a → (Integer, a)

scaleExp ∷ RealFrac a ⇒ a → (Integer, a)

data CompareResult a

Constructors

BELOW a Int

ABOVE a Int

MaxReached Int

taylorExpCmp ∷ RealFrac a ⇒ a → a → a → CompareResult a

Efficient way to compare the result of the Taylor expansion of the exponential function to a threshold value. Using error estimation one can stop early, once it's known the result will certainly be above or below the target value.